addMiddleware method
