+++ lib/collection.js
@@ -1,27 +1,11 @@
-var objectAssign = require('object-assign')
 var applyMiddlewares = require('./applyMiddlewares')
 
 module.exports = Collection
 
-var FIELDS_TO_CAST = ['operations', 'query', 'data', 'update']
-
-var DEFAULT_OPTIONS = {
-  castIds: true,
-  wrapNon$UpdateField: true,
-  middlewares: [
-    require('monk-plugin-query'),
-    require('monk-plugin-options'),
-    require('monk-plugin-cast-ids')(FIELDS_TO_CAST),
-    require('monk-plugin-fields'),
-    require('monk-plugin-handle-callback'),
-    require('monk-plugin-wait-for-connection')
-  ]
-}
-
 function Collection (manager, name, options) {
   this.manager = manager
   this.name = name
-  this.options = objectAssign({}, DEFAULT_OPTIONS, options || {})
+  this.options = options
 
   this.middlewares = this.options.middlewares || []
   delete this.options.middlewares
