+++ lib/manager.js
@@ -32,6 +32,22 @@ var STATE = {
   OPEN: 'open'
 }
 
+var FIELDS_TO_CAST = ['operations', 'query', 'data', 'update']
+
+var DEFAULT_OPTIONS = {
+  safe: true,
+  castIds: true,
+  wrapNon$UpdateField: true,
+  middlewares: [
+    require('monk-plugin-query'),
+    require('monk-plugin-options'),
+    require('monk-plugin-cast-ids')(FIELDS_TO_CAST),
+    require('monk-plugin-fields'),
+    require('monk-plugin-handle-callback'),
+    require('monk-plugin-wait-for-connection')
+  ]
+}
+
 /*
  * Module exports.
  */
@@ -64,7 +80,7 @@ function Manager (uri, opts, fn) {
 
   opts = opts || {}
 
-  this._collectionOptions = objectAssign({safe: true}, opts.collectionOptions || {})
+  this._collectionOptions = objectAssign({}, DEFAULT_OPTIONS, opts.collectionOptions || {})
   delete opts.collectionOptions
 
   if (Array.isArray(uri)) {
@@ -116,6 +132,7 @@ function Manager (uri, opts, fn) {
   this.collection = this.col = this.get = this.get.bind(this)
   this.oid = this.id
   this.setDefaultCollectionOptions = this.setDefaultCollectionOptions.bind(this)
+  this.addMiddleware = this.addMiddleware.bind(this)
 }
 
 /*
@@ -275,6 +292,16 @@ Manager.prototype.create = function (name, creationOptions, options) {
 
 Manager.prototype.setDefaultCollectionOptions = function (options) {
   this._collectionOptions = options
+}
+
+Manager.prototype.addMiddleware = function (middleware) {
+  if (!this._collectionOptions) {
+    this._collectionOptions = {}
+  }
+  if (!this._collectionOptions.middlewares) {
+    this._collectionOptions.middlewares = []
+  }
+  this._collectionOptions.middlewares.push(middleware)
 }
 
 Manager.prototype.id = function (str) {
